pyproject.toml
[project]
name = "amazonia-speech-gateway"
version = "0.1.0"
description = "Gateway para transcrição de áudio usando Whisper"
authors = [
    {name = "Peterson D.",email = "peterson.duarte@waproject.com.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "protobuf (==5.29.5)",
    "grpcio-tools (==1.71)",
    "grpcio (>=1.73.0,<2.0.0)",
    "grpcio-reflection (>=1.70.0,<2.0.0)",
    "opentelemetry-api (>=1.34.1,<2.0.0)",
    "opentelemetry-sdk (>=1.34.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.34.1,<2.0.0)",
    "opentelemetry-instrumentation-grpc (>=0.55b1,<0.56)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "oci (>=2.129.0,<3.0.0)",
    "boto3 (>=1.35.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [
    { include = "grpc_server" },
    { include = "core" },
    { include = "pkg" },
]


[tool.poetry.group.dev.dependencies]
mypy = "^1.16.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
